[FIX] (notification) 주석 추가 및 정리

+++ src/main/java/com/example/lastproject/domain/notification/service/NotificationServiceImpl.java
@@ -234,7 +234,9 @@ public class NotificationServiceImpl implements NotificationService {
     @Override
     @Transactional
     public void readNotification(Long notificationId, AuthUser authUser) {
-        Notification notification = verifyNotificationAccess(notificationId, authUser);
+        verifyNotificationAccess(notificationId, authUser);
+        Notification notification = notificationRepository.findById(notificationId)
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_NOTIFICATION));
         notification.read();
     }
 
@@ -246,7 +248,9 @@ public class NotificationServiceImpl implements NotificationService {
     @Override
     @Transactional
     public void deleteNotification(Long notificationId, AuthUser authUser) {
-        Notification notification = verifyNotificationAccess(notificationId, authUser);
+        verifyNotificationAccess(notificationId, authUser);
+        Notification notification = notificationRepository.findById(notificationId)
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_NOTIFICATION));
         notificationRepository.delete(notification);
     }
 
@@ -254,18 +258,15 @@ public class NotificationServiceImpl implements NotificationService {
      * 특정 ID의 알림을 조회하고 권한을 검증합니다.
      * @param notificationId 알림의 고유 ID
      * @param authUser 요청을 보낸 인증된 사용자 정보
-     * @return 알림 객체
      * @throws CustomException 해당 ID의 알림이 존재하지 않거나 접근 권한이 없을 경우 발생합니다.
      */
-    public Notification verifyNotificationAccess(Long notificationId, AuthUser authUser) {
+    public void verifyNotificationAccess(Long notificationId, AuthUser authUser) {
         Notification notification = notificationRepository.findById(notificationId)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_NOTIFICATION));
 
         if (!notification.getReceiver().getId().equals(authUser.getUserId())) {
             throw new CustomException(ErrorCode.UNAUTHORIZED_ACCESS);
         }
-
-        return notification;
     }
 
 }
